Factorize Logical interpreter
