+++ moses/atomese/interpreter/logical_interpreter.h
@@ -24,6 +24,10 @@
 #ifndef MOSES_LOGICAL_INTERPRETER_H
 #define MOSES_LOGICAL_INTERPRETER_H
 
+#include <opencog/atoms/base/Handle.h>
+#include <opencog/atoms/proto/LinkValue.h>
+#include <boost/iterator/zip_iterator.hpp>
+
 namespace opencog{ namespace atomese{
 
 /**
@@ -62,8 +66,8 @@ struct zip_or :
 
 	void operator()(const boost::tuple<const ProtoAtomPtr&, const ProtoAtomPtr&>& t)
 	{
-		if(HandleCast(t.get<0>())->get_type() == TRUE_LINK
-		   || HandleCast(t.get<1>())->get_type() == TRUE_LINK){
+		if((HandleCast(t.get<0>())->get_type() == TRUE_LINK)
+		   || (HandleCast(t.get<1>())->get_type() == TRUE_LINK)){
 			_result.push_back(ProtoAtomPtr(createLink(TRUE_LINK)));
 		}
 		else{
@@ -72,5 +76,76 @@ struct zip_or :
 	}
 };
 
+/**
+ * Compute logical_and.
+ * it expects the values of the LinkValues to be TRUE_LINK and FALSE_LINK
+ * @param LinkValuePtr&     p1
+ * @param LinkValuePtr&     p2
+ *
+ * @return                 LinkValue pointer containing the logical_and.
+ */
+LinkValuePtr logical_and(const LinkValuePtr& p1, const LinkValuePtr& p2){
+	std::vector<ProtoAtomPtr> p1_value = p1->value();
+	std::vector<ProtoAtomPtr> p2_value = p2->value();
+
+	zip_and _and = std::for_each(
+			boost::make_zip_iterator(
+					boost::make_tuple(p1_value.begin(), p2_value.begin())
+			),
+			boost::make_zip_iterator(
+					boost::make_tuple(p1_value.end(), p2_value.end())
+			),
+			zip_and()
+	);
+
+	std::vector<ProtoAtomPtr> _result = _and._result;
+	return LinkValuePtr(new LinkValue(_result));
+}
+
+/**
+ * Compute logical_or.
+ * it expects the values of the LinkValues to be TRUE_LINK and FALSE_LINK
+ * @param LinkValuePtr&     p1
+ * @param LinkValuePtr&     p2
+ *
+ * @return                 LinkValue pointer containing the logical_or.
+ */
+LinkValuePtr logical_or(const LinkValuePtr& p1, const LinkValuePtr& p2){
+	std::vector<ProtoAtomPtr> p1_value = p1->value();
+	std::vector<ProtoAtomPtr> p2_value = p2->value();
+
+	zip_or _or = std::for_each(
+			boost::make_zip_iterator(
+					boost::make_tuple(p1_value.begin(), p2_value.begin())
+			),
+			boost::make_zip_iterator(
+					boost::make_tuple(p1_value.end(), p2_value.end())
+			),
+			zip_or()
+	);
+
+	std::vector<ProtoAtomPtr> _result = _or._result;
+	return LinkValuePtr(new LinkValue(_result));
+}
+
+/**
+ * Compare Contents of two link_values.
+ * @param LinkValuePtr&     p1
+ * @param LinkValuePtr&     p2
+ *
+ * @return                 boolean of the comparision.
+ */
+bool logical_compare(const LinkValuePtr& p1, const LinkValuePtr& p2){
+	std::vector<ProtoAtomPtr> p1_value = p1->value();
+	std::vector<ProtoAtomPtr> p2_value = p2->value();
+
+	std::function<bool (const ProtoAtomPtr&, const ProtoAtomPtr&)> comparator = []
+			(const ProtoAtomPtr& left, const ProtoAtomPtr& right)
+	{
+		return HandleCast(left)->get_type()==HandleCast(right)->get_type();
+	};
+
+	return std::equal(p1_value.begin(), p1_value.end(), p2_value.begin(), comparator);
+}
 }}
 #endif //MOSES_LOGICAL_INTERPRETER_H
